Refactor lsm9ds1 IMU blocks to new module.
