+++ src/modules/editor.js
@@ -51,6 +51,7 @@ import './blockly/generators/propc/sensors/gps';
 import './blockly/generators/propc/sensors/hmc5883l_compass';
 import './blockly/generators/propc/sensors/joystick';
 import './blockly/generators/propc/sensors/lis3dh_accelerometer';
+import './blockly/generators/propc/sensors/lsm9ds1_imu';
 import './blockly/generators/propc/sensors/mma7455';
 import './blockly/generators/propc/sensors/mx2125';
 import './blockly/generators/propc/sensors/ping';
